+++ Runtime/PlayerLoopFramePacer.cs
@@ -129,49 +129,6 @@ namespace DTech.Pulse
 			}
 		}
 		
-		private static bool TryAppendSubSystem(ref PlayerLoopSystem root, Type parentType, PlayerLoopSystem system)
-		{
-			PlayerLoopSystem[] subSystems = root.subSystemList;
-			if (subSystems == null)
-			{
-				return false;
-			}
-
-			var copy = new PlayerLoopSystem[subSystems.Length];
-			Array.Copy(subSystems, copy, subSystems.Length);
-
-			for (int i = 0; i < copy.Length; i++)
-			{
-				if (copy[i].type == parentType)
-				{
-					PlayerLoopSystem[] children = copy[i].subSystemList;
-					int childrenCount = children?.Length ?? 0;
-					var newChildren = new PlayerLoopSystem[childrenCount + 1];
-					if (childrenCount > 0)
-					{
-						Array.Copy(children, newChildren, childrenCount);
-					}
-
-					newChildren[childrenCount] = system;
-					copy[i].subSystemList = newChildren;
-					root.subSystemList = copy;
-					return true;
-				}
-
-				PlayerLoopSystem child = copy[i];
-				if (!TryAppendSubSystem(ref child, parentType, system))
-				{
-					continue;
-				}
-
-				copy[i] = child;
-				root.subSystemList = copy;
-				return true;
-			}
-
-			return false;
-		}
-
 		private void TickFrame()
 		{
 			TaskCompletionSource<bool>[] completedWaits;
@@ -207,7 +164,7 @@ namespace DTech.Pulse
 				updateDelegate = TickFrame,
 			};
 
-			if (!TryAppendSubSystem(ref modifiedPlayerLoop, typeof(UnityEngine.PlayerLoop.Update), pacerSystem))
+			if (!PlayerLoopUtilities.TryAppendSubSystem(ref modifiedPlayerLoop, typeof(UnityEngine.PlayerLoop.Update), pacerSystem))
 			{
 				return false;
 			}